refactor: move config file to server directory
Makes more sense there because it’s the only place the file is needed.
