+++ server/config.js
@@ -1,5 +1,5 @@
 const config = {
-	APP_FRIENDLY_NAME: 'Template',
+	APP_FRIENDLY_NAME: 'fill-out--your-user-facing-app-name',
 	IS_PRODUCTION: process.env.NODE_ENV === 'production',
 	PORT: process.env.PORT || 3001
 };
